requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
annotated-types=0.7.0=pypi_0
anyio=4.8.0=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asttokens=3.0.0=pypi_0
async-lru=2.0.4=pypi_0
attrs=25.1.0=pypi_0
babel=2.17.0=pypi_0
beautifulsoup4=4.13.3=pypi_0
bleach=6.2.0=pypi_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.2.25=haa95532_0
certifi=2025.1.31=pypi_0
cffi=1.17.1=pypi_0
charset-normalizer=3.4.1=pypi_0
colorama=0.4.6=pypi_0
comm=0.2.2=pypi_0
contourpy=1.3.1=pypi_0
cycler=0.12.1=pypi_0
debugpy=1.8.12=pypi_0
decorator=5.2.1=pypi_0
defusedxml=0.7.1=pypi_0
distro=1.9.0=pypi_0
executing=2.2.0=pypi_0
expat=2.6.4=h8ddb27b_0
fastjsonschema=2.21.1=pypi_0
fonttools=4.56.0=pypi_0
fqdn=1.5.1=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.7=pypi_0
httpx=0.28.1=pypi_0
idna=3.10=pypi_0
ipykernel=6.29.5=pypi_0
ipython=8.32.0=pypi_0
ipywidgets=8.1.5=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.5=pypi_0
jiter=0.8.2=pypi_0
json5=0.10.0=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
jupyter=1.1.1=pypi_0
jupyter-client=8.6.3=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=5.7.2=pypi_0
jupyter-events=0.12.0=pypi_0
jupyter-lsp=2.2.5=pypi_0
jupyter-server=2.15.0=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyterlab=4.3.5=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.27.3=pypi_0
jupyterlab-widgets=3.0.13=pypi_0
kiwisolver=1.4.8=pypi_0
libffi=3.4.4=hd77b12b_1
libmpdec=4.0.0=h827c3e9_0
markupsafe=3.0.2=pypi_0
matplotlib=3.10.1=pypi_0
matplotlib-inline=0.1.7=pypi_0
mistune=3.1.2=pypi_0
nbclient=0.10.2=pypi_0
nbconvert=7.16.6=pypi_0
nbformat=5.10.4=pypi_0
nest-asyncio=1.6.0=pypi_0
notebook=7.3.2=pypi_0
notebook-shim=0.2.4=pypi_0
numpy=2.2.3=pypi_0
openai=1.64.0=pypi_0
openssl=3.0.15=h827c3e9_0
overrides=7.7.0=pypi_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.4=pypi_0
pillow=11.1.0=pypi_0
pip=25.0=py313haa95532_0
platformdirs=4.3.6=pypi_0
prometheus-client=0.21.1=pypi_0
prompt-toolkit=3.0.50=pypi_0
psutil=7.0.0=pypi_0
pure-eval=0.2.3=pypi_0
pycparser=2.22=pypi_0
pydantic=2.10.6=pypi_0
pydantic-core=2.27.2=pypi_0
pygments=2.19.1=pypi_0
pyparsing=3.2.1=pypi_0
python=3.13.2=hadb2040_100_cp313
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
python-json-logger=3.2.1=pypi_0
python_abi=3.13=0_cp313
pytz=2025.1=pypi_0
pywin32=308=pypi_0
pywinpty=2.0.15=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=26.2.1=pypi_0
referencing=0.36.2=pypi_0
requests=2.32.3=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rpds-py=0.23.1=pypi_0
send2trash=1.8.3=pypi_0
setuptools=75.8.0=py313haa95532_0
six=1.17.0=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.6=pypi_0
sqlite=3.45.3=h2bbff1b_0
stack-data=0.6.3=pypi_0
terminado=0.18.1=pypi_0
tinycss2=1.4.0=pypi_0
tk=8.6.14=h0416ee5_0
tornado=6.4.2=pypi_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pypi_0
types-python-dateutil=2.9.0.20241206=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2025.1=pypi_0
uri-template=1.3.0=pypi_0
urllib3=2.3.0=pypi_0
vc=14.42=haa95532_4
vs2015_runtime=14.42.34433=he0abc0d_4
wcwidth=0.2.13=pypi_0
webcolors=24.11.1=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.8.0=pypi_0
wheel=0.45.1=py313haa95532_0
widgetsnbextension=4.0.13=pypi_0
xz=5.6.4=h4754444_1
zlib=1.2.13=h8cc25b3_1
